feat: add CpuStat and ProcessStat classes for /proc/stat and /proc/[pid]/stat parsing
Introduce CpuStat and ProcessStat to encapsulate parsing of /proc/stat and /proc/[pid]/stat respectively. ProcPidStatIdx enum class keyed to the proc(5) ABI for /proc/[pid]/stat. Promote LinuxParser::UpTime from float to double to avoid precision loss in long-uptime calculations.

Add proc/[pid]/stat fixture files and fixture-based tests covering CPU aggregate field parsing and per-process time field extraction.

+++ include/linux_parser.h
@@ -21,7 +21,7 @@ namespace LinuxParser
 
   // System
   [[deprecated("Use MemInfo from 'meminfo.h' to retrieve, store and access this info instead.")]] float MemoryUtilization(const std::string &root = "/");
-  float UpTime(const std::string &root = "/");
+  double UpTime(const std::string &root = "/");
   std::vector<int> Pids(const std::string &root = "/");
   int TotalProcesses(const std::string &root = "/");
   int RunningProcesses();
